splits gossip Protocol out of cluster_info.rs
