+++ gossip/src/crds_gossip_push.rs
@@ -13,11 +13,12 @@
 
 use {
     crate::{
-        cluster_info::{Ping, CRDS_UNIQUE_PUBKEY_CAPACITY},
+        cluster_info::CRDS_UNIQUE_PUBKEY_CAPACITY,
         crds::{Crds, CrdsError, Cursor, GossipRoute},
         crds_gossip,
         crds_value::CrdsValue,
         ping_pong::PingCache,
+        protocol::Ping,
         push_active_set::PushActiveSet,
         received_cache::ReceivedCache,
     },
